[REFACTOR] interface와 imple 분리 및 폴더구조 변경
