+++ LionHeart-iOS/LionHeart-iOS/Scenes/Coordinator/ControllerableInterface/LoginViewControllerable.swift
@@ -1,0 +1,12 @@
+//
+//  LoginViewControllerable.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/15.
+//
+
+import UIKit
+
+protocol LoginViewControllerable where Self: UIViewController {
+    var coordinator: LoginNavigation? { get set }
+}
