+++ LionHeart-iOS/LionHeart-iOS/Scenes/Coordinator/ControllerableInterface/SplashViewControllerable.swift
@@ -1,0 +1,12 @@
+//
+//  SplashViewControllerable.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/15.
+//
+
+import UIKit
+
+protocol SplashViewControllerable where Self: UIViewController {
+    var coordinator: SplashNavigation? { get set }
+}
