+++ LionHeart-iOS/LionHeart-iOS/Scenes/Factory/FactoryImpl/TodayFactoryImpl.swift
@@ -1,19 +1,14 @@
 //
-//  TodayFactory.swift
+//  TodayFactoryImpl.swift
 //  LionHeart-iOS
 //
-//  Created by 황찬미 on 2023/10/15.
+//  Created by uiskim on 2023/10/15.
 //
 
 import UIKit
-
-protocol TodayFactory {
-    func makeTodayViewController() -> TodayViewControllerable
-}
 
 struct TodayFactoryImpl: TodayFactory {
     func makeTodayViewController() -> TodayViewControllerable {
         return TodayViewController(manager: TodayManagerImpl(articleService: ArticleServiceImpl(apiService: APIService())))
     }
 }
-
