+++ LionHeart-iOS/LionHeart-iOS/Scenes/Factory/FactoryInterface/AuthFactory.swift
@@ -1,0 +1,16 @@
+//
+//  AuthFactoryImpl.swift
+//  LionHeart-iOS
+//
+//  Created by 황찬미 on 2023/10/15.
+//
+
+import Foundation
+
+protocol AuthFactory {
+    func makeLoginViewController() -> LoginViewControllerable
+    func makeCompleteOnbardingViewController() -> CompleteOnbardingViewControllerable
+    func makeOnboardingViewController() -> OnboardingViewControllerable
+}
+
+
