+++ LionHeart-iOS/LionHeart-iOS/Scenes/Factory/FactoryInterface/SpalshFactory.swift
@@ -1,0 +1,14 @@
+//
+//  SpalshFactory.swift
+//  LionHeart-iOS
+//
+//  Created by 황찬미 on 2023/10/15.
+//
+
+import UIKit
+
+protocol SplashFactory {
+    func makeSplashViewController() -> SplashViewControllerable
+}
+
+
