+++ LionHeart-iOS/LionHeart-iOS/Scenes/Factory/FactoryInterface/TodayFactory.swift
@@ -1,0 +1,15 @@
+//
+//  TodayFactory.swift
+//  LionHeart-iOS
+//
+//  Created by 황찬미 on 2023/10/15.
+//
+
+import UIKit
+
+protocol TodayFactory {
+    func makeTodayViewController() -> TodayViewControllerable
+}
+
+
+
