+++ LionHeart-iOS/LionHeart-iOS/Scenes/Login/LoginViewController.swift
@@ -13,10 +13,6 @@ import SnapKit
 import KakaoSDKAuth
 import KakaoSDKUser
 
-protocol LoginViewControllerable where Self: UIViewController {
-    var coordinator: LoginNavigation? { get set }
-}
-
 final class LoginViewController: UIViewController, LoginViewControllerable {
     var userData: UserOnboardingModel?
