+++ LionHeart-iOS/LionHeart-iOS/Scenes/Onboarding/ViewControllers/CompleteOnbardingViewController.swift
@@ -10,11 +10,6 @@ import UIKit
 
 import SnapKit
 
-protocol CompleteOnbardingViewControllerable where Self: UIViewController {
-    var coordinator: CompleteOnbardingNavigation? { get set }
-    var userData: UserOnboardingModel? { get set }
-}
-
 final class CompleteOnbardingViewController: UIViewController, CompleteOnbardingViewControllerable {
     
     private enum SizeInspector {
