+++ LionHeart-iOS/LionHeart-iOS/Scenes/Today/ViewControllers/TodayViewController.swift
@@ -10,10 +10,6 @@ import UIKit
 
 import SnapKit
 
-protocol TodayViewControllerable where Self: UIViewController {
-    var coordinator: TodayNavigation? { get set }
-}
-
 final class TodayViewController: UIViewController, TodayViewControllerable {
     
     weak var coordinator: TodayNavigation?
